requirements.txt
chardet==3.0.4
Django==2.2.8
django-cors-headers==3.2.0
django-redis==4.10.0
djangorestframework==3.10.3
gunicorn==20.0.4
mysqlclient==1.4.6
pymongo==3.9.0
redis==3.3.11
requests==2.22.0
numpy==1.18.4
pdfkit==0.6.1
Pillow==7.1.2
django-debug-toolbar==3.1.1
